:recycle: DOP-3874 adds useCanonicalUrl custom hook for retrieving the canonical
